refactor: rename `goutils` to `fp`
